Validate store presence on orders
I need to make some behaviour dependent on the store an order is
associated to. Under normal circumstances, the order helper takes care
of that in 100% of cases. However, there's no presence validation on
the store, and for the feature I'm envisioning (Differing default taxation
depending on which store you access) I need every order to have a store.

The validation has to be on the `:order_id`, as just validating `:order`
will usually pass because `Spree::Store.default` returns a new, invalid
store if none is there.

This entails a lot of spec setup changes, mostly to make sure a store exists.