Quick return out of add_store_credit_payment
If we don't have a user, we don't need to be inside of
add_store_credit_payment.

If we have a user that doesn't have any credit, and our order doesn't
have any credit, we don't need to be inside of add_store_credit_payment.

To accomplish this we need to *change existing behavior* of
add_store_credit_payment to not add the error message given the previous
changes. I argue its previous implementation was overstepping its bounds
in assuming it should know when and how to add that error onto the
order.

Related to #908