Ensure every order has a store - Rake task and Changelog
Solidus 1.3 will not work well when it encounters an order with no
store set. This commit adds a changelog saying that and as well a
sample rake task that fixes up orders for relatively simple shops.

Also, it introduces a task `solidus:upgrade:one_point_three` which
can be hooked up with other data migration tasks we might need.