Reduce the surface area of Spree::Stock::Coordinator
Make 'shipments' the only public method of the Coordinator. It's the
only method called on Coordinator outside of the specs.

This was prompted by me looking at the Coordinator code while
reviewing some of the taxes work and thinking it might be nice to
refactor it a bit, which is easier with a reduced surface area.