Hide niche stock restriction feature for cleaner backed UI.
The backend interface has in the user profile a "Stock Locations"
configuration. It's not clear from the UI what this feature is at all
(my first assumption was it was some indicator of where a customer should
get products shipped from by default).

Looked into it and it seems to be a order fullfillment feature to
ensure folks that are fullfilling the order can only do so out of the
stock locations they are assigned (so a worker in Seattle cannnot
fullfill with inventory from Atlanta).

This seems very much a niche feature as most stores only have one location
and even stores that do have multiple locations may not have problems
fulfilling from the wrong location. I would argue this entire feature should
be an extension.

But going down that path of moduralizing that feature is probably not worth
the effort at this point so just extended it to be less prominent. Added
a new preference that will indicate if this configuration should be shown.
Since I hate to have yet another configuration I made the default intelligent
so that it looks to see if one of the RestrictedStock permission sets are
being used to determine the default.

If you are using a custom permission set and therefore the intelligent
default fails you can still manually indicate you are using this feature but
hopefully for stores using it the feature will just work as before and for
stores not using it, they get a cleaner UI.