Add bind-polyfill.js
Phantomjs 1.9.8 (still our recommended version) doesn't support
Function.prototype.bind. This adds a polyfill for it.

If we ever choose to only support phantomjs >= 2.0, this can be removed,
but we must first switch our CI and it would be polite to give our users
a great deal of notice, as this will affect any tests they have of the
backend.