Update spree_adjustments.finalized to be non null
When we migrated from state=open/closed to finalized=true/false in
d33254d we added `finalized` as a nullable column.  This breaks the
scopes that we added in that commit -- when finalized is null the
adjustments are neither `finalized` nor `not_finalized`.

I noticed this while trying to add some code and specs using the
`not_finalized` scope.

We could try to tweak the scopes but I think it'd be better to just
avoid the ternary logic altogether by disallowing nulls.

In the old code we had `state_machine :state, initial: :open` which
prevented this problem even though the `state` column itself was
nullable.