Improve Specs for Spree::Calculator::DefaultTax
Previously, there was no specs for when to refund VATs.

This adds a spec for turning a VAT into a discount when the
order's tax zone has nothing in common with the default or rates VAT zone.

Furthermore, this breaks out the specs for how to tax a taxable item
into a `shared_examples_for` block and tests it with line items,
shipments and shipping rates.