Better selected page detection for admin menu.
Previously if you used a custom URL you had to do something like this:

    <%= tab :leads, match_path: 'batch_communications/leads',
      url: main_app.admin_batch_communications_leads_path if can? :admin, Lead %>

Basically we are specifying the path twice. Once to indicate where we should
like to and once to indicate if the we are at that path. Generally if we link to
the path then we should also match that path. This patch allows the above to
work without the `match_path` option.

If you specify a `match_path` that will take priority. This just provides a nice
default to reduce the amount of boilerplate code.