Connect ConservativeLineItemPricer to LineItem
The pricer returns a Money object which has to align with the
currency and price of the line item. For the time being, the easiest here
is setting both (the validation will catch things).

This required adding a little bit of power to `Spree::Money` - namely,
`:to_d` and `:currency`.