Port current_currency helper module from spree_multi_domain
Previously, the `current_currency` helper would always return the
configured default currency for the Solidus installation. Solidus core
allows setting a default currency per `Spree::Store`, and this commit
adds a `current_currency` helper that respects that setting.

This does purposefully not include setting a currency via the session
(like the `spree_multi_currency` gem does. Solidus core does not allow
this, and it is much less complex to switch currencies with the configured
store.