Use new Controller Helper module "Pricing" in frontend and API
Before this commit, both API and Core defined a `current_currency`
helper. In the case of API, this helper was not tested, and in both
cases, the Helper only loaded the default currency. With a store model
in core that can hold a `Spree::Store`-wide default currency, this
method had to get a little more complex, and thus could not be defined in
two places anymore.

This commit makes use of the the new `Pricing` controller helper module
and adds it into the `Spree::StoreController` as well as the
`Spree::Api::BaseController`.