Simplify buggy tax rate migration
This migration did not work, and requires a call to `TaxRate#compute_amount`.

We could have fixed the "the does not work", but the `compute_amount` call requires
half of Solidus' object graph (including the `Spree::Calculator` STI table as well
as `Spree::Order`, `Spree::Shipment`, `Spree::TaxRate` - this is still not exhaustive).

We felt adding all of those into the migration was less than ideal, especially if your
local shipping rate was taxed using a different system than the default tax calculator.

The downside of this is ONLY that in the backend for old shipping orders, you will not be
shown the shipping rate tax (included or additional).

The reduced-in-complexity migration now just adds the `tax_rate_id` column on `Spree::ShippingRate`.