Remove Conservative Line Item pricer, use Variant pricer instead
The line item pricers have been a not-so-good idea. They over-engineer
behaviour of the line item itself while not playing well with
`pricing_options`. Replaced the class with an if statement.

Added some tests so that the pricing options will return a currency even if
the orders currency is nil.