Add breadcrumbs helpers
Currently breadcrumbs are created by piling content into the
"page_title" yield block.

It would be great if the breadcrumbs were somehow "meta" information,
provided by a controller action. This would require better organization
of the backend controllers. For now, I'd like to add some structure to
the breadcrumb information by separating the "levels" into individual
entries. This does two things for us:

1. It homogeonizes the way in which breadcrumbs are created and makes it
more clear how different pages are related. This could help us move them
around or develop a better systme for organizing the backend.
2. It gives us easier access to "the last breadcrumb" which is very
often "the actual title of the page". This then allows us to pick better
page titles, and maybe in the future *only* show the page title, and
drop the breadcrumbs altogether, in favour of a better navigation.