Rename PriceGenerator to VatPriceGenerator
This price generator is only necessary for shops with VAT.
This name change reflects this coupling. Also added an early return to
show that the thing does nothing when there are no VATs.

Rename Pricer to PriceSelector

"To price" can be understood as "setting a price" or "selecting a price"
"Selecting a price" is much closer to what we actually want here.